# Conflicts:
#	TravelWithPleasureHub/pom.xml
#	TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/controller/UserController.java
#	TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/service/UserService.java
#	TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/service/impl/UserServiceImpl.java
#	TravelWithPleasureHub/src/main/resources/application.properties

+++ TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/controller/MeetingController.java
@@ -73,7 +73,7 @@ public class MeetingController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<MeetingDTO> findMeetingById(@PathVariable long id) {
+    public ResponseEntity<MeetingDTO> findMeetingById(@PathVariable Integer id) {
         log.debug("REST request to get Meeting by id : {}", id);
         Optional<MeetingDTO> meetingDTO = meetingService.findById(id);
         return meetingDTO.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
